update the modules for all work
